requirements.txt
streamlit
google-genai
pypdf
plotly
st-gsheets-connection